+++ src/clarity_value/neon_encoder.rs
@@ -22,7 +22,7 @@ use clarity::vm::types::{CharType, PrincipalData, SequenceData, Value as Upstrea
 use neon::prelude::*;
 use stacks_codec::StacksMessageCodec;
 
-use crate::address::c32::c32_address;
+use crate::address::c32_address;
 use crate::hex::{encode_hex, encode_hex_no_prefix};
 
 /// Emit a single Clarity value as a JS object on `cur_obj`.
